Move data dependency into separate module
