+++ app/src/main/java/com/rsa/cleannewsapp/core/di/ApplicationModule.java
@@ -4,11 +4,11 @@ import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 
 import com.rsa.cleannewsapp.CleanNewsApplication;
+import com.rsa.cleannewsapp.core.data.remote.NewsApi;
+import com.rsa.cleannewsapp.core.data.remote.NewsService;
 import com.rsa.cleannewsapp.core.data.remote.RemoteServiceFactory;
+import com.rsa.cleannewsapp.core.data.repository.DefaultNewsArticleRepository;
 import com.rsa.cleannewsapp.core.domain.repository.NewsArticleRepository;
-import com.rsa.cleannewsapp.features.newsarticle.remote.NewsApi;
-import com.rsa.cleannewsapp.features.newsarticle.remote.NewsService;
-import com.rsa.cleannewsapp.features.newsarticle.repository.DefaultNewsArticleRepository;
 
 import android.content.Context;
