+++ data/build.gradle
@@ -1,0 +1,54 @@
+apply plugin: 'com.android.library'
+
+android {
+    compileSdkVersion 29
+    buildToolsVersion "29.0.1"
+
+
+    defaultConfig {
+        minSdkVersion 15
+        targetSdkVersion 29
+        versionCode 1
+        versionName "1.0"
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+            buildConfigField("String", "NEWS_API_KEY", '"3a8f068cf53c400d93fb90e9e1320b49"')
+            buildConfigField("String", "NEWS_API_DOMAIN", '"https://newsapi.org"')
+            buildConfigField("String", "NEWS_API_VERSION", '"v2"')
+        }
+        debug {
+            buildConfigField("String", "NEWS_API_KEY", '"3a8f068cf53c400d93fb90e9e1320b49"')
+            buildConfigField("String", "NEWS_API_DOMAIN", '"https://newsapi.org"')
+            buildConfigField("String", "NEWS_API_VERSION", '"v2"')
+        }
+    }
+
+}
+
+dependencies {
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation project(':domain')
+
+//    implementation 'androidx.appcompat:appcompat:1.0.2'
+//    testImplementation 'junit:junit:4.12'
+//    androidTestImplementation 'androidx.test:runner:1.2.0'
+//    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
+
+    implementation mainApplication.retrofit
+    implementation mainApplication.rxjavaAndroid
+    implementation mainApplication.rxjava
+    implementation mainApplication.retrofitRxAdapter
+    implementation mainApplication.retrofitGsonAdapter
+    implementation mainApplication.retrofitLogginInterceptor
+
+    testImplementation testingDependencies.junitExt
+    testImplementation testingDependencies.mockito
+    testImplementation testingDependencies.truth
+}
