+++ data/src/test/java/com/rsa/cleannewsapp/core/data/repository/DefaultNewsArticleRepositoryTest.java
@@ -1,12 +1,12 @@
-package com.rsa.cleannewsapp.features.newsarticle.repository;
+package com.rsa.cleannewsapp.core.data.repository;
 
-import com.rsa.cleannewsapp.features.newsarticle.remote.NewsService;
+import com.rsa.cleannewsapp.core.data.remote.NewsService;
 
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.runners.MockitoJUnitRunner;
 
 import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.MatcherAssert.assertThat;
