+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':domain'
+include ':app', ':domain', ':data'
